Added constant in  utils folder
